Remove unnecessary calls to operations
