+++ src/utils/hooks/assetsHooks.ts
@@ -146,7 +146,3 @@ export const useIsLoading = () => {
 export const useLastTimeUpdated = () => {
   return useAppSelector(state => state.assets.lastTimeUpdated);
 };
-
-export const useGetLatestOperations = () => {
-  return useAppSelector(state => state.assets.latestOperations);
-};
