+++ src/utils/redux/slices/assetsSlice.test.ts
@@ -24,7 +24,6 @@ describe("assetsSlice", () => {
       refetchTrigger: 0,
       lastTimeUpdated: null,
       isLoading: false,
-      latestOperations: [],
     });
   });
 
@@ -46,7 +45,6 @@ describe("assetsSlice", () => {
       refetchTrigger: 0,
       lastTimeUpdated: null,
       isLoading: false,
-      latestOperations: [],
     });
 
     store.dispatch(
@@ -72,7 +70,6 @@ describe("assetsSlice", () => {
       refetchTrigger: 0,
       lastTimeUpdated: null,
       isLoading: false,
-      latestOperations: [],
     });
   });
 
@@ -106,7 +103,6 @@ describe("assetsSlice", () => {
       refetchTrigger: 0,
       lastTimeUpdated: null,
       isLoading: false,
-      latestOperations: [],
     });
   });
 
@@ -134,7 +130,6 @@ describe("assetsSlice", () => {
       refetchTrigger: 0,
       lastTimeUpdated: null,
       isLoading: false,
-      latestOperations: [],
     });
   });
 
@@ -160,7 +155,6 @@ describe("assetsSlice", () => {
       refetchTrigger: 0,
       lastTimeUpdated: null,
       isLoading: false,
-      latestOperations: [],
     });
   });
